// lib/types.ts
export type Chain =
  | "bitcoin"
  | "ethereum"
  | "polygon"
  | "bsc"
  | "arbitrum"
  | "optimism"
  | "base"
  | "avalanche"
  | "solana"
  | "tron"